+++ docs/changes.rst
@@ -13,6 +13,7 @@ To be released.
 
  - Import :mod:`collections.abc` explicitly. [:issue:`398` by Stefan Naumann]
  - Fixed memory leak in :class:`~wand.image.HistogramDict`. [:issue:`397`]
+ - Added :meth:`Image.implode() <wand.image.BaseImage.implode>` method.
  - Added :meth:`Image.wave() <wand.image.BaseImage.wave>`.
  - Added :meth:`Image.white_threshold() <wand.image.BaseImage.white_threshold>`.
  - Added :meth:`Image.vignette() <wand.image.BaseImage.vignette>`.